requirements.txt
django
djangorestframework
pillow
django-cors-headers
ipykernel
